Revert "Add PaywallsV2 header component (#6548)"
This reverts commit a30910679f10c5603a6ddfe14251abdf5507fa52.
