+++ RevenueCatUI/Templates/V2/Components/Root/RootViewModel.swift
@@ -25,20 +25,17 @@ class RootViewModel {
         let parentZStack: PaywallComponent.StackComponent?
     }
 
-    let headerViewModel: HeaderComponentViewModel?
     let stackViewModel: StackComponentViewModel
     let stickyFooterViewModel: StickyFooterComponentViewModel?
     let firstItemIgnoresSafeAreaInfo: FirstItemShouldIgnoreSafeAreaInfo?
     let localizationProvider: LocalizationProvider
 
     init(
-        headerViewModel: HeaderComponentViewModel?,
         stackViewModel: StackComponentViewModel,
         stickyFooterViewModel: StickyFooterComponentViewModel?,
         firstItemIgnoresSafeAreaInfo: FirstItemShouldIgnoreSafeAreaInfo?,
         localizationProvider: LocalizationProvider
     ) {
-        self.headerViewModel = headerViewModel
         self.stackViewModel = stackViewModel
         self.stickyFooterViewModel = stickyFooterViewModel
         self.firstItemIgnoresSafeAreaInfo = firstItemIgnoresSafeAreaInfo
